finish video decoder V1

+++ VideoDecoder/nv12_to_rgb_resize.cuh
@@ -3,8 +3,8 @@
 #include "../common.h"
 
 void nv12_to_rgb_resize(
-	f32* out_resized_rgb_frames,
-	f32 const* const batched_yuv_frames,
+	u8* out_resized_rgb_frames,
+	u8 const* const batched_yuv_frames,
 	usize batch_size,
 	usize nv12_frame_size_in_bytes,
 	i32 input_width,
